unit test token name validation regex
