+++ auth/test/pytest.ini
@@ -1,0 +1,11 @@
+[pytest]
+asyncio_mode = auto
+addopts = --strict-markers
+markers =
+    unchecked_allocator: tests that use the unchecked allocator
+    asyncio: test files that use asyncio
+filterwarnings =
+    error
+    ignore::UserWarning
+    ignore::DeprecationWarning
+xfail_strict=true
